pyproject.toml
[tool.poetry]
name = "pai"
version = "0.1.0"
description = ""
authors = ["endaaman <buhibuhidog@gmail.com>"]

[tool.poetry.scripts]
client = "entrypoint:client"
server = "entrypoint:server"

[tool.poetry.dependencies]
python = ">=3.6.1, <4"
numpy = "^1.18.2"
gbulb = { git = "https://github.com/endaaman/gbulb.git", branch = "master" }
pillow = "^7.1.1"
pycairo = "^1.19.1"
asyncio-glib = "^0.1"
aiohttp = "^3.6.2"
websockets = "^8.1"
flask = "^1.1.2"
tornado = "^6.0.4"

[tool.poetry.dev-dependencies]
pylint = "^2.4.4"
jedi = "^0.16.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

